Fix VC++ compiler warning
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11301 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/laserdisc/OggReader.cc
@@ -627,7 +627,7 @@ const AudioFragment* OggReader::getAudio(unsigned sample)
 
 		// read more if we're at the end of the list
 		if (it == audioList.end()) {
-			unsigned size = audioList.size();
+			size_t size = audioList.size();
 			while (size == audioList.size()) {
 				if (!nextPacket()) {
 					return NULL;
